+++ CHANGELOG.md
@@ -11,6 +11,10 @@ since 0.4.1, and this project adheres to [Semantic Versioning](https://semver.or
 
 - Use the archived backports for Debian bullseye
   ([#1213](https://github.com/freedomofpress/dangerzone/issues/1213))
+- Install `numpy` 2.0.0 in Python 3.9 envs, and use more recent `numpy` in
+  Python environments >= 3.10, to avoid some compatibility issues with Python
+  3.13
+  ([#1206](https://github.com/freedomofpress/dangerzone/issues/1206))
 
 ## [0.9.1](https://github.com/freedomofpress/dangerzone/compare/v0.9.1...0.9.0)